Changed the way posts are rendered
